+++ spec/lumberjack/entry_formatter_spec.rb
@@ -4,18 +4,18 @@ require "spec_helper"
 
 RSpec.describe Lumberjack::EntryFormatter do
   describe "building formatters" do
-    it "starts with an empty message formatter and no attribute formatter" do
+    it "starts with an empty message formatter and attribute formatter" do
       entry_formatter = Lumberjack::EntryFormatter.new
       expect(entry_formatter.message_formatter).to_not be_nil
-      expect(entry_formatter.attribute_formatter).to be_nil
+      expect(entry_formatter.attribute_formatter).to_not be_nil
       obj = Object.new
       expect(entry_formatter.message_formatter.format(obj)).to equal(obj)
+      expect(entry_formatter.attribute_formatter).to be_empty
     end
 
     it "uses the default formatter if message_formatter is :default" do
       entry_formatter = Lumberjack::EntryFormatter.new(message_formatter: :default)
       expect(entry_formatter.message_formatter).to_not be_nil
-      expect(entry_formatter.attribute_formatter).to be_nil
       obj = Object.new
       expect(entry_formatter.message_formatter.format(obj)).to eq(obj.inspect)
     end
@@ -38,46 +38,48 @@ RSpec.describe Lumberjack::EntryFormatter do
     it "can add new attribute formatters in a chain" do
       entry_formatter = Lumberjack::EntryFormatter.new
       formatter = lambda {}
-      expect(entry_formatter.attributes { add(Object, formatter) }).to eq(entry_formatter)
+      expect(entry_formatter.add_attribute_class(Object, formatter)).to eq(entry_formatter)
       expect(entry_formatter.attribute_formatter).to_not be_empty
     end
 
     it "can remove attribute formatters in a chain" do
       entry_formatter = Lumberjack::EntryFormatter.new
       formatter = lambda {}
-      entry_formatter.attributes { add(Object, formatter) }
-      expect(entry_formatter.attributes { remove(Object) }).to eq(entry_formatter)
+      entry_formatter.add_attribute_class(Object, formatter)
+      expect(entry_formatter.remove_attribute_class(Object)).to eq(entry_formatter)
       expect(entry_formatter.attribute_formatter).to be_empty
     end
 
+    it "can set the attribute default formatter in a chain" do
+      entry_formatter = Lumberjack::EntryFormatter.new
+      entry_formatter.default_attribute_format { |obj| obj.to_s.upcase }
+      expect(entry_formatter.attribute_formatter.format(foo: "bar")).to eq({"foo" => "BAR"})
+    end
+
     it "build a formatter with a build block" do
-      entry_formatter = Lumberjack::EntryFormatter.build do
-        add(String) { |obj| obj.to_s.upcase }
-        attributes do
-          add("status") { |obj| "[#{obj}]" }
-        end
+      entry_formatter = Lumberjack::EntryFormatter.build do |config|
+        config.add(String) { |obj| obj.to_s.upcase }
+        config.add_attribute("status") { |obj| "[#{obj}]" }
+        config.add_attribute_class(Array) { |obj| obj.join("|") }
       end
 
       expect(entry_formatter.message_formatter.format("foobar")).to eq("FOOBAR")
       expect(entry_formatter.attribute_formatter.format("status" => "new")).to eq({"status" => "[new]"})
+      expect(entry_formatter.attribute_formatter.format("tags" => ["a", "b", "c"])).to eq({"tags" => "a|b|c"})
     end
   end
 
   describe "#include" do
     it "merges the formats from the formatter" do
-      formatter_1 = Lumberjack::EntryFormatter.build do
-        add(String) { |obj| obj.to_s.upcase }
-        attributes do
-          add("status") { |obj| "[#{obj}]" }
-          add("foo") { |obj| "foo:#{obj}" }
-        end
+      formatter_1 = Lumberjack::EntryFormatter.build do |config|
+        config.add(String) { |obj| obj.to_s.upcase }
+        config.add_attribute("status") { |obj| "[#{obj}]" }
+        config.add_attribute("foo") { |obj| "foo:#{obj}" }
       end
 
-      formatter_2 = Lumberjack::EntryFormatter.build do
-        add(String) { |obj| obj.to_s.downcase }
-        attributes do
-          add("foo") { |obj| "(#{obj})" }
-        end
+      formatter_2 = Lumberjack::EntryFormatter.build do |config|
+        config.add(String) { |obj| obj.to_s.downcase }
+        config.add_attribute("foo") { |obj| "(#{obj})" }
       end
 
       expect(formatter_2.include(formatter_1)).to eq formatter_2
@@ -91,19 +93,15 @@ RSpec.describe Lumberjack::EntryFormatter do
 
   describe "#prepend" do
     it "prepends the formats from the formatter" do
-      formatter_1 = Lumberjack::EntryFormatter.build do
-        add(String) { |obj| obj.to_s.upcase }
-        attributes do
-          add("status") { |obj| "[#{obj}]" }
-          add("foo") { |obj| "foo:#{obj}" }
-        end
+      formatter_1 = Lumberjack::EntryFormatter.build do |config|
+        config.add(String) { |obj| obj.to_s.upcase }
+        config.add_attribute("status") { |obj| "[#{obj}]" }
+        config.add_attribute("foo") { |obj| "foo:#{obj}" }
       end
 
-      formatter_2 = Lumberjack::EntryFormatter.build do
-        add(String) { |obj| obj.to_s.downcase }
-        attributes do
-          add("foo") { |obj| "(#{obj})" }
-        end
+      formatter_2 = Lumberjack::EntryFormatter.build do |config|
+        config.add(String) { |obj| obj.to_s.downcase }
+        config.add_attribute("foo") { |obj| "(#{obj})" }
       end
 
       expect(formatter_2.prepend(formatter_1)).to eq formatter_2
@@ -145,7 +143,7 @@ RSpec.describe Lumberjack::EntryFormatter do
     end
 
     it "applies the attribute formatter to the attributes" do
-      entry_formatter.attributes { add("foo") { |obj| "Foo: #{obj}" } }
+      entry_formatter.add_attribute("foo") { |obj| "Foo: #{obj}" }
       message, attributes = entry_formatter.format("foobar", {"foo" => "bar"})
       expect(message).to eq("foobar")
       expect(attributes).to eq({"foo" => "Foo: bar"})
@@ -165,7 +163,7 @@ RSpec.describe Lumberjack::EntryFormatter do
     end
 
     it "handles nil attributes" do
-      entry_formatter.attributes { add("foo") { |obj| "Foo: #{obj}" } }
+      entry_formatter.add_attribute("foo") { |obj| "Foo: #{obj}" }
       message, attributes = entry_formatter.format("foobar", nil)
       expect(message).to eq("foobar")
       expect(attributes).to be_nil