+++ spec/lumberjack_spec.rb
@@ -86,8 +86,8 @@ RSpec.describe Lumberjack do
   end
 
   it "can build a formatter" do
-    formatter = Lumberjack.build_formatter do
-      add(Integer) { |i| i * 2 }
+    formatter = Lumberjack.build_formatter do |config|
+      config.add(Integer) { |i| i * 2 }
     end
     expect(formatter).to be_a(Lumberjack::EntryFormatter)
     expect(formatter.format(12, nil).first).to eq(24)